requirements.txt
# Pinned on 2019-04-13
aniso8601==6.0.0
attrs==19.1.0
Click==7.0
environ-config==18.2.0
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
